refactor: model을 소켓과 restApi 공통으로 관리하기 위해 분리
