+++ src/network/model/common/response.model.ts
@@ -1,0 +1,7 @@
+interface CommonResponse<T> {
+  result: T;
+  resultCode: number;
+  resultMsg: string;
+}
+
+export type { CommonResponse };
