+++ src/network/model/library/get/getAssetList.model.ts
@@ -1,0 +1,22 @@
+import { DomainType } from '@/features/assetLibrary/constants/domain';
+
+interface RequestGetAssetList {
+  domain: string;
+  majorCategory: string;
+  minorCategory: string;
+  page: number;
+}
+
+interface AssetInfo {
+  id: string;
+  name: string;
+  filename: string;
+  type: string;
+  domain: DomainType;
+  createdAt: string;
+  author: string;
+}
+
+type ResponseGetAssetList = AssetInfo[];
+
+export type { RequestGetAssetList, ResponseGetAssetList };
