+++ src/network/model/project/delete/deleteMxProject.model.ts
