--- src/network/webSocket/services/model/commonResponse.model.ts
@@ -1,7 +1,0 @@
-interface ResponseDto<T> {
-  result: T;
-  resultCode: number;
-  resultMsg: string;
-}
-
-export type { ResponseDto };
